+++ cli/src/pcluster/dcv/connect.py
@@ -14,7 +14,7 @@ import subprocess as sub
 import time
 import webbrowser
 
-from api.pcluster_api import FullClusterInfo, PclusterApi
+from api.pcluster_api import ClusterInfo, PclusterApi
 from pcluster.constants import PCLUSTER_ISSUES_LINK
 from pcluster.dcv.utils import DCV_CONNECT_SCRIPT
 from pcluster.utils import error, get_region
@@ -39,7 +39,7 @@ def dcv_connect(args):
     :param args: pcluster cli arguments.
     """
     result = PclusterApi().describe_cluster(cluster_name=args.cluster_name, region=get_region())
-    if isinstance(result, FullClusterInfo):
+    if isinstance(result, ClusterInfo):
         # Prepare ssh command to execute in the head node instance
         cmd = 'ssh {CFN_USER}@{HEAD_NODE_IP} {KEY} "{REMOTE_COMMAND} /home/{CFN_USER}"'.format(
             CFN_USER=result.user,